fix(causal): worker loader packed unexpected module
