+++ packages/rath-client/src/utils/user.ts
@@ -1,0 +1,67 @@
+import md5 from "crypto-js/md5";
+import Cookies from 'js-cookie';
+
+
+export enum IAVATAR_TYPES {
+    gravatar = 'gravatar',
+    default = 'default'
+}
+
+// 此模块与 Rath 其他部分单向解耦，
+// 不要在这里添加对其他模块的依赖（包括 types, constants 和 utils），
+// 或将应当置于此模块的内容拆移到外部。
+// 一部分语句（如下方的 赋值表达式[右值 = 成员表达式]）
+// 如果侵入到其他（被 worker 依赖的）模块有可能被经过 loader 错误打包，
+// 导致所有 Web Workers 无法注册。
+
+export const AVATAR_IMG_LIST: string[] = new Array(18)
+    .fill(0)
+    .map((_, i) => `avatar-B-${`${i + 1}`.padStart(2, '0')}.png`);
+
+export const DEFAULT_AVATAR_URL_PREFIX = 'https://foghorn-assets.s3.ap-northeast-1.amazonaws.com/avatar/';
+
+const AUTH_COOKIE_KEY = 'userAuthed';
+
+export function setLoginCookie(userName: string) {
+    Cookies.set(AUTH_COOKIE_KEY, userName, { expires: 1 });
+}
+export function checkLoginCookie() {
+    const userName = Cookies.get(AUTH_COOKIE_KEY);
+    if (userName && userName !== '') {
+        return userName;
+    }
+    return null;
+}
+
+export function clearLoginCookie() {
+    Cookies.remove(AUTH_COOKIE_KEY);
+}
+
+export function getServerUrl(path: string) {
+    const baseURL = new URL(window.location.href);
+    const DATA_SERVER_URL =
+        baseURL.searchParams.get('main_service') || localStorage.getItem('main_service') || window.location.href;
+    // const devSpecURL = new URL(w|| window.location.href)
+    const url = new URL(DATA_SERVER_URL);
+    url.pathname = path;
+    return url.toString();
+}
+
+export function getDefaultAvatarURL(imgKey: string, size: 'small' | 'large' = 'large') {
+    return `${DEFAULT_AVATAR_URL_PREFIX}${size}/${imgKey}`;
+}
+
+type AvatarProps = {
+    avatarType: IAVATAR_TYPES;
+    avatarKey: string;
+    size: 'small' | 'large';
+    email: string;
+}
+
+export function getAvatarURL(props: AvatarProps) {
+    if (props.avatarType === IAVATAR_TYPES.default) {
+        return `${DEFAULT_AVATAR_URL_PREFIX}${props.size}/${props.avatarKey}`
+    } else {
+        return `https://www.gravatar.com/avatar/${md5(props.email.toLowerCase()).toString()}`;
+    }
+}
